pyproject.toml
[build-system]
requires = [
    "setuptools >= 64"
]
build-backend = "setuptools.build_meta"

[project]
name = "catapult-util"
description = "Catapult Utilities"
readme = "README.rst"
authors = [{name = "Lukas Moors", email = "me@lukasmoors.com"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["locking", "eventhub"]
requires-python = ">= 3.9"
dependencies = [
    "pymongo > 4.0",
    "azure-eventhub >= 5.15.0"
]
dynamic = ["version"]
